// src/types/api/setting.ts
export type Setting = {
  category1: {
    name: string;
    age: number;
    tall: number;
    weight: number;
    address: string;
    test1: number;
    test2: number;
  };
  category2: {
    name: string;
    age: number;
    tall: number;
    weight: number;
  };
  category3: {
    name: string;
    age: number;
    tall: number;
    weight: number;
    address: string;
    test1: number;
    test2: number;
  };
  category4: {
    name: string;
    age: number;
    tall: number;
    weight: number;
    address: string;
    test1: number;
    test2: number;
  };
};